requirements.txt
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
loguru==0.6.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
